extract populateWeatherData logic into apis module
